+++ src/components/UserMenu.tsx
@@ -2,8 +2,9 @@ import Image from "next/future/image";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { signOut } from "next-auth/react";
-import { useState } from "react";
+import { useState, useRef } from "react";
 import { useSWRConfig } from "swr";
+import useOnClickOutside from "use-onclickoutside";
 import HamburgerIcon from "../../public/static/images/hamburger-50.png";
 import { Confirm } from "./Confirm";
 import { deleteRide } from "../hooks";
@@ -21,6 +22,7 @@ export const UserMenu = ({ role, rideId }: MenuProps) => {
   const isLeader = role && ["ADMIN", "LEADER"].includes(role);
   const router = useRouter();
   const { mutate } = useSWRConfig();
+  const ref = useRef(null);
 
   const toggleMenu = () => {
     setShow(!show);
@@ -58,8 +60,10 @@ export const UserMenu = ({ role, rideId }: MenuProps) => {
 
   const confirmDelete = () => setShowConfirm(true);
 
+  useOnClickOutside(ref, closeMenu);
+
   return (
-    <div className="relative">
+    <div ref={ref} className="relative">
       <div className="h-10 cursor-pointer rounded p-1 hover:bg-neutral-200">
         <button type="button" onClick={toggleMenu} onKeyDown={toggleMenu}>
           <Image